Add files via upload
Project: Saving_Superstore.

Implementing my skills as a consultant to review the superstore’s operations and increase its profitability to avoid bankruptcy.

GOAL: To review multiple area's of interest in order to ascertain what could be causing the losses before the company files for bankruptcy. 

We've broken it down into the following categories of review:

Profits and Losses ,

Top 3 and Bottom 3 product categories ,

Returns (Client based and Product based) ,

Marketing trends (throughout each Month and across each State) ,

Profits vs Averages.

We've discovered that the state responsible for the the most profits, is incurring losses ;which would account for the Bottom 3 product categories being returned at an increasing rate , while the Top 3 are selling fairly easily and with consistency. So it leads us to believe the issue may be with quality and overall satisfaction.

Within the Bottom 3 Products, we've ascertained that Bookcases, Supplies, and Tables are doing the worst. While the Top 3 Products are Accessories, Copiers, and Phones.

From the data presented there are 4 customers who are averaging between 11 and 14 returned items (which is way more than any other customer).

Sales Revenue within this state can also be attributed to the marketing campaigns being launched. 

CONCLUSION: The state that raked in the most revenue and allowed the Top 3 Categories to thrive, while also maintaining a steady and consistent marketing campaign for the Top 3 Categories is - California.